+++ src/Pim/Bundle/EnrichBundle/Form/Subscriber/FilterLocaleSpecificValueSubscriber.php
@@ -10,7 +10,7 @@ use Symfony\Component\Form\FormEvent;
  * Filter locale specific value subscriber to remove value available in only a set of locales
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 class FilterLocaleSpecificValueSubscriber implements EventSubscriberInterface