+++ src/Pim/Bundle/BaseConnectorBundle/Validator/Import/ProductImportValidator.php
@@ -2,13 +2,14 @@
 
 namespace Pim\Bundle\BaseConnectorBundle\Validator\Import;
 
-use Pim\Bundle\BaseConnectorBundle\Exception\DuplicateProductValueException;
+use Pim\Bundle\CatalogBundle\Manager\ProductManager;
 use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Validator\ConstraintGuesserInterface;
 use Pim\Bundle\TransformBundle\Transformer\ColumnInfo\ColumnInfoInterface;
 use Pim\Bundle\TransformBundle\Transformer\ProductTransformer;
+use Pim\Bundle\BaseConnectorBundle\Exception\DuplicateProductValueException;
 use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\ConstraintViolationListInterface;
 use Symfony\Component\Validator\ValidatorInterface;
@@ -30,31 +31,39 @@ class ProductImportValidator extends ImportValidator
     /**
      * @var array
      */
-    protected $constraints = array();
+    protected $constraints = [];
 
     /**
      * @var array
      */
-    protected $uniqueValues = array();
+    protected $uniqueValues = [];
+
+    /**
+     * @var ProductManager
+     */
+    protected $productManager;
 
     /**
      * Constructor
      *
      * @param ValidatorInterface         $validator
      * @param ConstraintGuesserInterface $constraintGuesser
+     * @param ProductManager             $productManager
      */
     public function __construct(
         ValidatorInterface $validator,
-        ConstraintGuesserInterface $constraintGuesser
+        ConstraintGuesserInterface $constraintGuesser,
+        ProductManager $productManager = null
     ) {
         parent::__construct($validator);
         $this->constraintGuesser = $constraintGuesser;
+        $this->productManager    = $productManager;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function validate($entity, array $columnsInfo, array $data, array $errors = array())
+    public function validate($entity, array $columnsInfo, array $data, array $errors = [])
     {
         if (null === $this->getIdentifier($columnsInfo, $entity)) {
             $notBlank = new NotBlank();
@@ -62,7 +71,11 @@ class ProductImportValidator extends ImportValidator
         }
 
         $this->checkIdentifier($entity, $columnsInfo, $data);
-        $this->checkUniqueValues($entity, $columnsInfo, $data);
+        try {
+            $this->checkUniqueValues($entity, $columnsInfo, $data);
+        } catch (DuplicateProductValueException $e) {
+            $errors[$data['sku']] = [[$e->getMessage()]];
+        }
 
         foreach ($columnsInfo as $columnInfo) {
             if ($columnInfo->getAttribute()) {
@@ -110,6 +123,9 @@ class ProductImportValidator extends ImportValidator
                     $code      = $value->getAttribute()->getCode();
                     $valueData = (string) $value;
                     if ($valueData !== '') {
+                        if ($this->productManager !== null && $this->productManager->valueExists($value)) {
+                            throw new DuplicateProductValueException($code, $valueData, $data);
+                        }
                         $this->uniqueValues[$code] =
                             isset($this->uniqueValues[$code]) ? $this->uniqueValues[$code] : [];