+++ src/PimEnterprise/Bundle/VersioningBundle/Denormalizer/Flat/ProductValue/MetricDenormalizer.php
@@ -12,12 +12,14 @@
 namespace PimEnterprise\Bundle\VersioningBundle\Denormalizer\Flat\ProductValue;
 
 use Pim\Bundle\CatalogBundle\Factory\MetricFactory;
+use Symfony\Component\OptionsResolver\OptionsResolverInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * Metric flat denormalizer used for attribute types:
  * - pim_catalog_metric
  *
- * @author    Romain Monceau <romain@akeneo.com>
+ * @author Romain Monceau <romain@akeneo.com>
  */
 class MetricDenormalizer extends AbstractValueDenormalizer
 {
@@ -40,9 +42,15 @@ class MetricDenormalizer extends AbstractValueDenormalizer
      */
     public function denormalize($data, $class, $format = null, array $context = array())
     {
-        $value = $context['value'];
+        if ($data === null || $data === '') {
+            return null;
+        }
 
-        $data = $data === '' ? null : $data;
+        $resolver = new OptionsResolver();
+        $this->configContext($resolver);
+        $context = $resolver->resolve($context);
+
+        $value = $context['value'];
 
         if (null === $metric = $value->getMetric()) {
             $metric = $this->factory->createMetric($value->getAttribute()->getMetricFamily());
@@ -52,5 +60,14 @@ class MetricDenormalizer extends AbstractValueDenormalizer
         }
 
         return $metric;
+    }
+
+    /**
+     * Define context requirements
+     * @param OptionsResolverInterface $resolver
+     */
+    protected function configContext(OptionsResolverInterface $resolver)
+    {
+        $resolver->setRequired(['value']);
     }
 }