+++ CHANGELOG.md
@@ -67,12 +67,12 @@
 
 ## Bug fixes
 - PIM-3332: Fix incompatibility with overriden category due to usage of ParamConverter in ProductController
-- PIM-3309: Fix check on product value uniqueness
 
 # 1.2.x
 
 ## Bug fixes
 - PIM-3298 : Fix issue with locale specific property of an attribute when edit and mass edit
+- PIM-3309: Fix check on product value uniqueness
 - PIM-3229: Fix values for simple and multi select attributes with missing translations not being displayed in the grid
 
 # 1.2.11 (2014-10-31)