+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Bug fixes
 - PIM-3331: Fix draft creation when saving values for newly added attributes for the first time
-- PIM-3351: Test data value in PricesDenormalizer to avoid creating empty ProductPrices.
+- PIM-3351: Test data value in PricesDenormalizer to avoid creating empty ProductPrices
+- PIM-3301: Test data value in DateTimeDenormalizer to avoid reverted date to be set on current day
 
 # 1.0.4 (2014-10-31)